Factor out request and multipart parsing
